requirements.txt
gspread
oauth2client
fpdf2
yagmail
pandas
dotenv
flake8
black
python-telegram-bot
pdf2image
pillow